media: i2c: maxim-serdes: max9296a: refactor PHY configuration
Also add support for mapping PHY lanes.

MAX9296A does not support opperating in single PHY mode, all
PHYs are used in pairs of 2, PHY 0 + PHY 1 and PHY 2 + PHY 3.
PHY 1 and PHY 2 are masters.

Since it does not support opperating in single PHY mode, all
lane configurations can be set for two PHYs at the same time.

PHY 0 outputs on lanes 2 and 3 by default, while PHY 1 outputs
on lanes 0 and 1. Clock lane is on PHY 1, the master PHY.

PHY 2 outputs on lines 0 and 1 by default, while PHY 3 outputs
on lanes 2 and 3. Clock lane is on PHY 2, the mater PHY.

DPLL is set on the master PHY.

Pipe remaps are set on the master PHY.